罗跃权 12.7 订单主表和详细表dao层

+++ entity/src/main/java/com/cn/tenmall/entity/WxTabCategoryReportEntity.java
@@ -18,5 +18,5 @@ public class WxTabCategoryReportEntity {
     private Integer categoryId3;
     private Date countDate;
     private Integer num;
-    private Integer money;
+    private Double money;
 }
